pyproject.toml
[project]
name = "trusted-agent-engine"
version = "2.0.0"
description = "Trusted AI Agent Governance Engine"
authors = [
    { name = "Yuangs AI Team" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "pydantic>=2.5.3",
    "json-logic>=0.6.3",
    "rich>=13.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.7",
    "httpx>=0.26.0",
    "cryptography>=42.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0"
]

[project.scripts]
trusted-engine = "trusted_agent_engine.cli.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["trusted_agent_engine*"]
